+++ src/components/content/ImageContent.jsx
@@ -3,38 +3,58 @@ import PropTypes, { ContentType } from '../prop-types';
 import BaseContent from './BaseContent';
 import { PRE_CALCULATED_RATIO } from '../../constants/ContentConstants';
 import EventBus, { Events } from '../../event';
+import { ViewType } from '../../constants/SettingConstants';
+import Connector from '../../service/connector';
 
-class ImageContent extends BaseContent {
+export default class ImageContent extends BaseContent {
   imageRef = React.createRef();
+  wrapperRef = React.createRef();
 
   constructor(props) {
     super(props);
 
+    this.state = {
+      imageInScreen: false,
+    };
+
     this.imageOnErrorHandler = this.imageOnErrorHandler.bind(this);
     this.imageOnLoadHandler = this.imageOnLoadHandler.bind(this);
+    this.imageInScreenHandler = this.imageInScreenHandler.bind(this);
   }
 
   componentDidMount() {
-    const { isCalculated, contentFooter, content } = this.props;
-    if (!isCalculated) {
-      EventBus.emit(Events.CALCULATE_CONTENT, {
-        index: content.index,
-        contentNode: this.imageRef.current,
-        contentFooterNode: contentFooter,
-        ratio: content.ratio,
-      });
+    super.componentDidMount();
+    const {
+      content,
+      isContentLoaded,
+    } = this.props;
+    if (!isContentLoaded) {
+
+      if ('IntersectionObserver' in window) {
+        const wrapper = this.wrapperRef.current;
+        const { columnsInPage } = Connector.setting.getSetting();
+        this.observer = new IntersectionObserver((entries) => {
+          if (entries.some(({ isIntersecting, intersectionRatio }) => isIntersecting || intersectionRatio > 0)) {
+            this.imageInScreenHandler();
+            this.observer.unobserve(wrapper);
+            this.observer = null;
+          }
+        }, { threshold: 0, rootMargin: `${200 * columnsInPage}%` });
+        this.observer.observe(wrapper);
+      } else {
+        this.timeout = setTimeout(this.imageInScreenHandler, (content.index - 1) * 200);
+      }
     }
   }
 
-  componentDidUpdate() {
-    const { isCalculated, contentFooter, content } = this.props;
-    if (!isCalculated) {
-      EventBus.emit(Events.CALCULATE_CONTENT, {
-        index: content.index,
-        contentNode: this.imageRef.current,
-        contentFooterNode: contentFooter,
-        ratio: content.ratio,
-      });
+  componentWillUnmount() {
+    if (this.observer) {
+      this.observer.unobserve(this.wrapperRef.current);
+      this.observer = null;
+    }
+    if (this.timeout) {
+      clearTimeout(this.timeout);
+      this.timeout = null;
     }
   }
 
@@ -57,6 +77,10 @@ class ImageContent extends BaseContent {
     }
   }
 
+  imageInScreenHandler() {
+    this.setState({ imageInScreen: true });
+  }
+
   renderImage() {
     const { src } = this.props;
     const { isContentOnError } = this.props.content;
@@ -67,8 +91,10 @@ class ImageContent extends BaseContent {
         </div>
       );
     }
+
     return (
       <img
+        className="img"
         ref={this.imageRef}
         src={src}
         alt=""
@@ -80,14 +106,23 @@ class ImageContent extends BaseContent {
 
   render() {
     const { contentFooter } = this.props;
-    const { isContentLoaded } = this.props.content;
+    const { isContentLoaded, ratio } = this.props.content;
+    const { imageInScreen } = this.state;
+    const { viewType } = Connector.setting.getSetting();
+    const style = {};
+    const renderImage = imageInScreen || isContentLoaded;
+
+    if (viewType === ViewType.SCROLL) {
+      style.paddingBottom = ratio === PRE_CALCULATED_RATIO ? '140%' : `${ratio * 100}%`;
+    }
     return (
       <section
-        ref={this.props.forwardedRef}
-        className={`comic_page ${isContentLoaded ? 'loaded' : ''} ${contentFooter ? 'has_content_footer' : ''}`}
+        ref={this.wrapperRef}
+        className={`image_container ${isContentLoaded ? 'loaded' : ''} ${contentFooter ? 'has_content_footer' : ''}`}
+        style={style}
       >
-        {this.renderImage()}
-        {contentFooter}
+        {renderImage && this.renderImage()}
+        {renderImage && contentFooter}
       </section>
     );
   }
@@ -95,15 +130,11 @@ class ImageContent extends BaseContent {
 
 ImageContent.defaultProps = {
   contentFooter: null,
-  forwardedRef: React.createRef(),
 };
 
 ImageContent.propTypes = {
   src: PropTypes.string,
   content: ContentType.isRequired,
   contentFooter: PropTypes.node,
-  forwardedRef: PropTypes.any,
   isCalculated: PropTypes.bool,
 };
-
-export default React.forwardRef((props, ref) => <ImageContent forwardedRef={ref} {...props} />);