+++ src/components/styled/StyledContent.jsx
@@ -38,9 +38,11 @@ const StyledHtmlContent = () => `
          local("Tahoma");
   }
 
-  font-size: ${Connector.setting.getFontSize(true)} !important;
-  line-height: ${Connector.setting.getLineHeight(true)} !important;
-  font-family: ${Connector.setting.getFont()} !important;
+  * {
+    font-size: ${Connector.setting.getFontSize(true)} !important;
+    line-height: ${Connector.setting.getLineHeight(true)} !important;
+    font-family: ${Connector.setting.getFont()} !important;
+  }
   
   img {
     max-width: 100%;
@@ -96,10 +98,12 @@ export const StyledImageScrollContent = StyledBaseContent.extend`
   ${StyledScrollContent}
   margin: 0 auto;
   .content_container {
+    position: relative;
     margin: 0 auto;
     width: ${() => Connector.setting.getContentWidth(1, true)};
-    img {
+    .img {
       width: 100%;
+      position: absolute;
     }
   }
 `;
@@ -111,15 +115,15 @@ export const StyledImagePageContent = StyledBaseContent.extend`
   margin: 0 auto;
   .content_container {
     &.two_images_in_page {
-      .comic_page {
+      .image_container {
         &:nth-child(odd) { img { margin-right: 0; } }
         &:nth-child(even) { img { margin-left: 0; } }
       }
     } 
     
-    .comic_page {
+    .image_container {
       height: 100%;
-      img {
+      .img {
         width: auto; height: auto;
         max-width: 100%; max-height: 100%;
         top: 50%;