+++ src/service/CurrentService.js
@@ -1,4 +1,4 @@
-import { timer, NEVER } from 'rxjs';
+import { timer, NEVER, merge } from 'rxjs';
 import {
   tap,
   map,
@@ -6,6 +6,7 @@ import {
   debounce,
   distinctUntilChanged,
   catchError,
+  mergeMap,
 } from 'rxjs/operators';
 import BaseService from './BaseService';
 import EventBus, { Events } from '../event';
@@ -24,7 +25,7 @@ class CurrentService extends BaseService {
     super.load();
 
     this.connectEvents(this.onCalculationInvalidated.bind(this), Events.CALCULATION_INVALIDATED);
-    this.connectEvents(this.onCalculated.bind(this), Events.CALCULATION_UPDATED);
+    this.connectEvents(this.onCalculated.bind(this), Events.CALCULATION_UPDATED, Events.CALCULATIONS_SET);
     this.connectEvents(this.onCalculationCompleted.bind(this), Events.CALCULATION_COMPLETED);
     this.connectEvents(this.onCurrentOffsetUpdated.bind(this), Events.UPDATE_CURRENT_OFFSET);
     this.connectEvents(this.onScrolled.bind(this), Events.SCROLL);
@@ -116,12 +117,21 @@ class CurrentService extends BaseService {
     });
   }
 
-  onCalculated(calculateUpdate$) {
-    return calculateUpdate$.pipe(
-      filter(() => !this._isOffsetRestored),
+  onCalculated(calculateUpdate$, calculationsSet$) {
+    return merge(
+      calculateUpdate$.pipe(
+        filter(() => !this._isOffsetRestored),
+      ),
+      calculationsSet$.pipe(
+        mergeMap(() => timer()),
+      ),
     ).subscribe(() => {
       if (this._restoreCurrentOffset() !== null) {
         this._isOffsetRestored = true;
+      }
+      if (!Connector.calculations.isReadyToRead()) {
+        Connector.calculations.setReadyToRead(true);
+        EventBus.emit(Events.READY_TO_READ);
       }
     });
   }